requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
fastapi==0.114.0
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
protobuf==5.28.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
python-multipart==0.0.9
requests==2.32.3
setuptools==74.1.2
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
validators==0.33.0
weaviate-client==4.7.1